requirements.txt
instructor==1.3.2
openai==1.33.0
pyautogen==0.2.28
pydantic==2.7.3
typing_extensions==4.12.2
